+++ src/config/netsendo.php
@@ -99,13 +99,13 @@ return [
 
     'plugins' => [
         'wordpress' => [
-            'version' => '1.0.0',
+            'version' => '1.1.0',
             'download_url' => '/plugins/wordpress/netsendo-wordpress.zip',
             'min_wp_version' => '5.8',
             'min_php_version' => '7.4',
         ],
         'woocommerce' => [
-            'version' => '1.0.0',
+            'version' => '1.1.0',
             'download_url' => '/plugins/woocommerce/netsendo-woocommerce.zip',
             'min_wp_version' => '5.8',
             'min_wc_version' => '5.0',